Split bdd test files
